Fix incorrect FCM connection logic
Prior to this, when running the Example app for the first time, the
following error was being printed in the console:

Unable to connect to FCM. Error Domain=com.google.fcm Code=501 "(null)"

After accepting the notifications permission, the following error was
being printed in the console:

Unable to connect to FCM. Error Domain=com.google.fcm Code=2001 "(null)"

The first error means that the app tried to connect to FCM while there
was no Firebase registration token available. It now checks for the
current token and exits early if it's null.

The second error means that the app tried to connect to FCM while
another connection is already available. It now disconnects from any
existing connection prior to trying to connect again.