Temporarily disable flaky firestore/testApplyFilter test (#877)
The testApplyFilter test assumes that there are a fixed number of
restaurants with the category Deli, and verifies that applying a filter
returns the right number of restaurants.

Unfortunately, the actual data behind this is random, and depends
entirely on the current state of the data in a shared project. If anyone
adds additional restaurants it's possible to add/remove Delis because
the client only loads the first 50 restaurants.

Tracked in b/147740878.